Prevent user from creating series named 'Unknown'
Since 'Unknown' is used in other fields to determine when that field does not have a value, the word 'Unknown' is now reserved and cannot be used as a name. Input_series() and edit_series use a list of reserved words to complete this task, making it easy to add other reserved words in the future.