Add EditSeries dialog to GUI
Displays the EditSeries dialog box when a series is selected and the ui_mainwindow.edit_series_button is clicked. The window retrieves the database rowid for the currently selected series and populates a table with that series's information, where the user can edit its properties, save those edits, or cancel editing.

Currently the dialog can only display info; in the future it will allow all series properties to be changed.