Add MangaTrackerAddWindow() class
When the 'add' button is clicked in the main window, a dialog opens with a table similar to the edit window, where the user can fill in information for a new series.

When finished, add_series will validate the contents of the new series, and add it to the database if valid.