Complete add_series() function
When each cell is edited, the validate_cells() function triggers and checks whether the input is valid, marking the background red if it is not. The add_series function checks that no cells are red, verifies the name is good, then attempts to add the series to the database. If it fails, the add series window stays open.

In the future, the window should show a tooltip stating what's wrong if the add_series() function fails.