Clarify requirements for changing database name
Previously, the program checked if either the new file did not already exist, or is_database() returned true for the new database file. Since is_database will return true for a non-existent file regardless, the check for os.path.exists() was unnecessary. Additionally, when the user inputs the current database filename, the CLI now correctly states that the database name was not changed.